Update State To Display Elapsed and Duration
I propose to change the state text to include the remain variable in addition to the total time. This code change also drops setting the presence starttimestamp as it's redundant. The purpose of this is to make it easier to identify how long someone has been watching/playing something in mpv.